Refactor Common code out of PyConfHoardDatastore
